Refactored after adding BackendPlanAdapter
