+++ sandbox/libs/analytics-framework/src/main/java/org/opensearch/analytics/spi/AnalyticsSearchBackendPlugin.java
@@ -8,8 +8,6 @@
 
 package org.opensearch.analytics.spi;
 
-import java.util.List;
-
 /**
  * SPI extension point for backend query engine plugins.
  *
@@ -72,12 +70,4 @@ public interface AnalyticsSearchBackendPlugin {
         return null;
     }
 
-    /**
-     * Returns RexNode transformers this backend requires during the marking phase.
-     * Transformers run on leaf predicates before backend capability matching.
-     * Empty list means no transformations needed.
-     */
-    default List<RexNodeTransformer> getRexTransformers() {
-        return List.of();
-    }
 }
